pyproject.toml
[project]
name = "ai-psychometricist"
version = "0.2.0"
description = "Multi-method conversational psychometric assessment — a scientific feasibility study"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4,<1.0",
    "langchain-openai>=0.3,<1.0",
    "langchain-core>=0.3,<1.0",
    "neo4j>=5.0,<6.0",
    "pydantic>=2.0,<3.0",
    "python-dotenv>=1.0,<2.0",
    "scipy>=1.14,<2.0",
    "numpy>=1.26,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.9.0",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "httpx>=0.27",
]
web = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.scripts]
interview = "src.main:main"
self-report = "src.evaluation.self_report:main"
compare = "src.evaluation.compare:main"
seed-graph = "src.graph.seed:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".venv", "data/sessions"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "B", "UP", "SIM"]

[tool.ruff.lint.per-file-ignores]
# (no per-file ignores needed)

[tool.mypy]
python_version = "3.11"
pretty = true
show_error_codes = true
ignore_missing_imports = true
warn_unused_ignores = true
follow_imports = "skip"
explicit_package_bases = true
files = ["src", "web", "tests"]
